backend

+++ src/controllers/todo/todo.controler.js
@@ -2,6 +2,7 @@ import { asyncHandler } from "../../middleware/asynnc-handler.middleware.js";
 import { Todo } from "../../models/todo/todo.model.js";
 import { HttpException } from "../../utils/http.exception.js";
 import { StatusCodes,ReasonPhrases }  from 'http-status-codes'
+import mongoose from 'mongoose';
 
 export const todoadd =  asyncHandler(async ( req, res) => {
     const {title, desc, image} = req.body;
@@ -47,9 +48,30 @@ export const get_all = async ( req, res ) => {
 
     res.status(200).json({ success: true, data })
 };
-export const delet = async ( req, res ) => {
-    const { id } = req.params;
-    await Todo.findByIdAndDelete(id);
+// export const delet = async ( req, res ) => {
+//     const { id } = req.params;
+//     await Todo.findByIdAndDelete(id);
     
-    res.status(200).json({ success: true, data, msg:"successfull file" },)
+//     res.status(200).json({ success: true, data, msg:"successfull file" },)
+// };
+const { ObjectId } = mongoose.Types;  // ObjectId ni olish
+export const delet = async (req, res) => {
+  const { id } = req.params;
+
+  // ID formatini tekshirish
+  if (!ObjectId.isValid(id)) {
+    return res.status(400).json({ success: false, msg: "Invalid ID format" });
+  }
+
+  // Todo ni bazadan o'chirish
+  const todo = await Todo.findByIdAndDelete(id);
+
+  if (!todo) {
+    return res.status(404).json({ success: false, msg: "Todo topilmadi" });
+  }
+
+  // O'chirilgan todo bilan javob yuborish
+  res.status(200).json({ success: true, data: todo, msg: "Todo muvaffaqiyatli o'chirildi" });
 };
+
+
